Wire compute room shields with rust
