+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/crypto/CryptoSessionInfoProvider.kt
@@ -17,13 +17,21 @@
 package org.matrix.android.sdk.internal.crypto
 
 import com.zhuinden.monarchy.Monarchy
+import org.matrix.android.sdk.api.crypto.RoomEncryptionTrustLevel
 import org.matrix.android.sdk.api.session.events.model.EventType
 import org.matrix.android.sdk.internal.database.model.EventEntity
 import org.matrix.android.sdk.internal.database.model.EventEntityFields
+import org.matrix.android.sdk.internal.database.model.RoomMemberSummaryEntity
+import org.matrix.android.sdk.internal.database.model.RoomMemberSummaryEntityFields
+import org.matrix.android.sdk.internal.database.model.RoomSummaryEntity
+import org.matrix.android.sdk.internal.database.query.where
 import org.matrix.android.sdk.internal.database.query.whereType
 import org.matrix.android.sdk.internal.di.SessionDatabase
+import org.matrix.android.sdk.internal.di.UserId
 import org.matrix.android.sdk.internal.session.room.membership.RoomMemberHelper
 import org.matrix.android.sdk.internal.util.fetchCopied
+import org.matrix.android.sdk.internal.util.logLimit
+import timber.log.Timber
 import javax.inject.Inject
 
 /**
@@ -31,7 +39,8 @@ import javax.inject.Inject
  * in the session DB, this class encapsulate this functionality
  */
 internal class CryptoSessionInfoProvider @Inject constructor(
-        @SessionDatabase private val monarchy: Monarchy
+        @SessionDatabase private val monarchy: Monarchy,
+        @UserId private val myUserId: String
 ) {
 
     fun isRoomEncrypted(roomId: String): Boolean {
@@ -47,7 +56,7 @@ internal class CryptoSessionInfoProvider @Inject constructor(
     /**
      * @param allActive if true return joined as well as invited, if false, only joined
      */
-     fun getRoomUserIds(roomId: String, allActive: Boolean): List<String> {
+    fun getRoomUserIds(roomId: String, allActive: Boolean): List<String> {
         var userIds: List<String> = emptyList()
         monarchy.doWithRealm { realm ->
             userIds = if (allActive) {
@@ -57,5 +66,42 @@ internal class CryptoSessionInfoProvider @Inject constructor(
             }
         }
         return userIds
+    }
+
+    fun getUserListForShieldComputation(roomId: String): List<String> {
+        var userIds: List<String> = emptyList()
+        monarchy.doWithRealm { realm ->
+            userIds = RoomMemberHelper(realm, roomId).getActiveRoomMemberIds()
+        }
+        var isDirect = false
+        monarchy.doWithRealm { realm ->
+            isDirect = RoomSummaryEntity.where(realm, roomId = roomId).findFirst()?.isDirect == true
+        }
+
+        return if (isDirect || userIds.size <= 2) {
+            userIds.filter { it != myUserId }
+        } else {
+            userIds
+        }
+    }
+
+    fun getRoomsWhereUsersAreParticipating(userList: List<String>): List<String> {
+        var roomIds: List<String>? = null
+        monarchy.doWithRealm { sessionRealm ->
+            roomIds = sessionRealm.where(RoomMemberSummaryEntity::class.java)
+                    .`in`(RoomMemberSummaryEntityFields.USER_ID, userList.toTypedArray())
+                    .distinct(RoomMemberSummaryEntityFields.ROOM_ID)
+                    .findAll()
+                    .map { it.roomId }
+                    .also { Timber.d("## CrossSigning -  ... impacted rooms ${it.logLimit()}") }
+        }
+        return roomIds.orEmpty()
+    }
+
+    fun updateShieldForRoom(roomId: String, shield: RoomEncryptionTrustLevel) {
+        monarchy.writeAsync { realm ->
+            val summary = RoomSummaryEntity.where(realm, roomId = roomId).findFirst()
+            summary?.roomEncryptionTrustLevel = shield
+        }
     }
 }
