+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/crypto/model/rest/KeysClaimResponse.kt
@@ -24,8 +24,8 @@ import com.squareup.moshi.JsonClass
  */
 @JsonClass(generateAdapter = true)
 internal data class KeysClaimResponse(
-        /// If any remote homeservers could not be reached, they are recorded here.
-        /// The names of the properties are the names of the unreachable servers.
+        // / If any remote homeservers could not be reached, they are recorded here.
+        // / The names of the properties are the names of the unreachable servers.
         @Json(name = "failures")
         val failures: Map<String, Any>,
